Talk To Me™
Talk To Me™ is an application under development at a very early stage of
production that is supposed to show how jFrame can be used to be a
Question and Answer software.

This software is meant to be used as an example that java is capable of
answering what is given.
Startup:
- You can choose a name by typing it into the chat bar.
- The software will remember your name.

Commands (currently case-sensitive):
/help - shows all commands
/restart - soft reset of software
/bye - exits software
/namechange <name> - changes your name
/clearchat - clears all chat
UP ARROW - previous entered line will be back in chat bar
DOWN ARROW - clears chat bar
"Talk To Me" - enables smart Talk To Me™ Bot

Talk To Me™ Smart Bot (currently case-sensitive, you must activate the
bot for it to reply back)
"hello" - replies back with a greeting
"hello, my name is <name>" - replies back with the bot's name and a
greeting

Made by InfernoDragon0